requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
qrcode==7.4.2
opencv-python==4.8.0.76
face_recognition==1.3.0
numpy==1.26.0
pandas==2.2.0
openpyxl==3.1.2
PyJWT==2.8.0
werkzeug==2.2.3
bcrypt==4.1.2
filelock==3.13.1
WTForms==3.1.1
Pillow==10.3.0

# Production WSGI Servers
waitress==2.1.2  # For Windows (recommended)
gunicorn==21.2.0  # For Linux/Unix

# Gemini API
google-generativeai==0.3.2